@@ -19,17 +19,18 @@
         }
     ],
     "require": {
-        "php": ">=8.0",
-        "ext-pdo": "*",
-        "squirrelphp/debug": "^2.0",
-        "doctrine/dbal": "^3.0"
+        "php": ">=8.2",
+        "squirrelphp/connection": "^0.2",
+        "squirrelphp/debug": "^2.0"
     },
     "require-dev": {
-        "bamarni/composer-bin-plugin": "^1.8",
-        "captainhook/plugin-composer": "^5.0",
-        "phpunit/phpunit": "^9.0",
+        "captainhook/captainhook-phar": "^5.0",
+        "captainhook/hook-installer": "^1.0",
+        "phpunit/phpunit": "^11.2",
         "mockery/mockery": "^1.0",
-        "squirrelphp/types": "^1.0"
+        "squirrelphp/types": "^1.0",
+        "symfony/finder": "^7.0",
+        "symfony/process": "^7.0"
     },
     "suggest": {
         "squirrelphp/queries-bundle": "Symfony integration of squirrelphp/queries - automatic assembling of decorated connections",
