@@ -29,7 +29,7 @@ class UpdateEntries implements BuilderInterface
     private bool $confirmNoWhere = false;
 
     public function __construct(
-        private DBInterface $db,
+        private readonly DBInterface $db,
     ) {
     }
 
