@@ -1,17 +1,17 @@
 <?php
 
-namespace Squirrel\Queries\Doctrine;
+namespace Squirrel\Queries\DB;
 
 use Squirrel\Debug\Debug;
 use Squirrel\Queries\DBInterface;
 use Squirrel\Queries\Exception\DBInvalidOptionException;
 
 /**
- * DB SQLite implementation using Doctrine DBAL with custom upsert functionality
+ * DB SQLite implementation with custom upsert functionality
  *
  * This is equal to the Postgres version, as SQLite uses the same syntax
  */
-class DBSQLiteImplementation extends DBPostgreSQLImplementation
+class SQLiteImplementation extends PostgreSQLImplementation
 {
     private ?float $sqliteVersion = null;
 
@@ -27,10 +27,10 @@ class DBSQLiteImplementation extends DBPostgreSQLImplementation
         if ($this->sqliteVersion === null) {
             $connection = $this->getConnection();
 
-            $statement = $connection->prepare('select sqlite_version() AS "v"');
-            $statementResult = $statement->executeQuery();
-            $result = $statementResult->fetchAssociative();
-            $statementResult->free();
+            $statement = $connection->prepareQuery('select sqlite_version() AS "v"');
+            $connection->executeQuery($statement);
+            $result = $connection->fetchOne($statement);
+            $connection->freeResults($statement);
 
             if (!isset($result['v'])) {
                 throw Debug::createException(
