@@ -2,46 +2,27 @@
 
 namespace Squirrel\Queries\Tests;
 
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Result;
-use Doctrine\DBAL\Statement;
-use Hamcrest\Core\IsEqual;
 use Mockery\MockInterface;
-use Squirrel\Queries\Doctrine\DBMySQLImplementation;
+use Squirrel\Connection\ConnectionInterface;
+use Squirrel\Queries\DB\MySQLImplementation;
 use Squirrel\Queries\Exception\DBInvalidOptionException;
 
 /**
  * Test custom MySQL implementation parts
  */
-class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
+class MySQLImplementationTest extends \PHPUnit\Framework\TestCase
 {
-    private DBMySQLImplementation $db;
-    /** @var Connection&MockInterface */
-    private $connection;
+    use SharedFunctionalityTrait;
 
-    /**
-     * Prepare common aspects of all tests
-     */
+    private MySQLImplementation $db;
+    private ConnectionInterface&MockInterface $connection;
+
     protected function setUp(): void
     {
-        // Mock of the Doctrine Connection class
-        $this->connection = \Mockery::mock(Connection::class);
+        $this->connection = \Mockery::mock(ConnectionInterface::class);
         $this->connection->shouldReceive('quoteIdentifier')->andReturnUsing([$this, 'quoteIdentifier']);
 
-        // MySQL implementation class
-        $this->db = new DBMySQLImplementation($this->connection);
-    }
-
-    private function bindValues(MockInterface $statement, array $vars): void
-    {
-        $varCounter = 1;
-
-        foreach ($vars as $var) {
-            $statement
-                ->shouldReceive('bindValue')
-                ->once()
-                ->with(IsEqual::equalTo($varCounter++), IsEqual::equalTo($var), \PDO::PARAM_STR);
-        }
+        $this->db = new MySQLImplementation($this->connection);
     }
 
     /**
@@ -49,7 +30,6 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
      */
     public function testUpsert(): void
     {
-        // SQL query which should be generated by the implementation
         $sql = 'INSERT INTO ' . $this->quoteIdentifier('example.example') .
             ' (' .
             $this->quoteIdentifier('id') . ',' .
@@ -73,31 +53,8 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
             1,
         ];
 
-        // Doctrine statement and result
-        $statement = \Mockery::mock(Statement::class);
-        $statementResult = \Mockery::mock(Result::class);
+        $this->prepareForQueryWithVars($sql, $vars);
 
-        $this->bindValues($statement, $vars);
-
-        $statement
-            ->shouldReceive('executeQuery')
-            ->once()
-            ->withNoArgs()
-            ->andReturn($statementResult);
-
-        // SQL query should be received by "prepare"
-        $this->connection
-            ->shouldReceive('prepare')
-            ->once()
-            ->with(IsEqual::equalTo($sql))
-            ->andReturn($statement);
-
-        // Close result set
-        $statementResult
-            ->shouldReceive('free')
-            ->once();
-
-        // Test the upsert
         $this->db->insertOrUpdate('example.example', [
             'id' => 5,
             'id2' => 6,
@@ -113,33 +70,10 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Recreated quoteIdentifier function as a standin for the Doctrine one
-     *
-     * @param string $identifier
-     * @return string
-     */
-    public function quoteIdentifier(string $identifier): string
-    {
-        if (strpos($identifier, ".") !== false) {
-            $parts = array_map(
-                function ($p) {
-                    return '"' . str_replace('"', '""', $p) . '"';
-                },
-                explode(".", $identifier),
-            );
-
-            return implode(".", $parts);
-        }
-
-        return '"' . str_replace('"', '""', $identifier) . '"';
-    }
-
-    /**
      * Test upsert with an explicit update part
      */
     public function testUpsertCustomUpdate(): void
     {
-        // SQL query which should be generated by the implementation
         $sql = 'INSERT INTO ' . $this->quoteIdentifier('example.example') . ' (' .
             $this->quoteIdentifier('id') . ',' .
             $this->quoteIdentifier('id2') . ',' .
@@ -164,31 +98,8 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
             'laaaast',
         ];
 
-        // Doctrine statement and result
-        $statement = \Mockery::mock(Statement::class);
-        $statementResult = \Mockery::mock(Result::class);
+        $this->prepareForQueryWithVars($sql, $vars);
 
-        $this->bindValues($statement, $vars);
-
-        $statement
-            ->shouldReceive('executeQuery')
-            ->once()
-            ->withNoArgs()
-            ->andReturn($statementResult);
-
-        // SQL query should be received by "prepare"
-        $this->connection
-            ->shouldReceive('prepare')
-            ->once()
-            ->with(IsEqual::equalTo($sql))
-            ->andReturn($statement);
-
-        // Close result set
-        $statementResult
-            ->shouldReceive('free')
-            ->once();
-
-        // Test the upsert
         $this->db->insertOrUpdate('example.example', [
             'id' => 5,
             'id2' => 6,
@@ -214,7 +125,6 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
      */
     public function testUpsertCustomUpdateWithVars(): void
     {
-        // SQL query which should be generated by the implementation
         $sql = 'INSERT INTO ' . $this->quoteIdentifier('example.example') . ' (' .
             $this->quoteIdentifier('id') . ',' .
             $this->quoteIdentifier('id2') . ',' .
@@ -241,31 +151,8 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
             'laaaast',
         ];
 
-        // Doctrine statement and result
-        $statement = \Mockery::mock(Statement::class);
-        $statementResult = \Mockery::mock(Result::class);
+        $this->prepareForQueryWithVars($sql, $vars);
 
-        $this->bindValues($statement, $vars);
-
-        $statement
-            ->shouldReceive('executeQuery')
-            ->once()
-            ->withNoArgs()
-            ->andReturn($statementResult);
-
-        // SQL query should be received by "prepare"
-        $this->connection
-            ->shouldReceive('prepare')
-            ->once()
-            ->with(IsEqual::equalTo($sql))
-            ->andReturn($statement);
-
-        // Close result set
-        $statementResult
-            ->shouldReceive('free')
-            ->once();
-
-        // Test the upsert
         $this->db->insertOrUpdate('example.example', [
             'id' => 5,
             'id2' => 6,
@@ -292,7 +179,6 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
      */
     public function testUpsertNoUpdateRows(): void
     {
-        // SQL query which should be generated by the implementation
         $sql = 'INSERT INTO ' . $this->quoteIdentifier('example.example') . ' (' .
             $this->quoteIdentifier('id') . ',' .
             $this->quoteIdentifier('id2') .
@@ -305,31 +191,8 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
             6,
         ];
 
-        // Doctrine statement and result
-        $statement = \Mockery::mock(Statement::class);
-        $statementResult = \Mockery::mock(Result::class);
+        $this->prepareForQueryWithVars($sql, $vars);
 
-        $this->bindValues($statement, $vars);
-
-        $statement
-            ->shouldReceive('executeQuery')
-            ->once()
-            ->withNoArgs()
-            ->andReturn($statementResult);
-
-        // SQL query should be received by "prepare"
-        $this->connection
-            ->shouldReceive('prepare')
-            ->once()
-            ->with(IsEqual::equalTo($sql))
-            ->andReturn($statement);
-
-        // Close result set
-        $statementResult
-            ->shouldReceive('free')
-            ->once();
-
-        // Test the upsert
         $this->db->insertOrUpdate('example.example', [
             'id' => 5,
             'id2' => 6,
@@ -343,10 +206,8 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
 
     public function testUpsertInvalidOptionNoTableName(): void
     {
-        // Expect an InvalidOptions exception
         $this->expectException(DBInvalidOptionException::class);
 
-        // Try it with the invalid option
         $this->db->insertOrUpdate('', [
             'dada' => 5,
             'fieldname' => 'rowvalue',
@@ -355,10 +216,8 @@ class DoctrineMySQLImplementationTest extends \PHPUnit\Framework\TestCase
 
     public function testUpsertInvalidOptionNoRow(): void
     {
-        // Expect an InvalidOptions exception
         $this->expectException(DBInvalidOptionException::class);
 
-        // Try it with the invalid option
         $this->db->insertOrUpdate('table');
     }
 }
